requirements.txt
amqp==5.0.6
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.1.0
Django==3.2.8
django-filter==21.1
django-redis==5.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
idna==3.2
ipdb==0.13.9
ipython==7.28.0
jedi==0.18.0
kombu==5.1.0
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
Pygments==2.10.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
